feat: brand items are revised to have a hyperlink, have a link to show trains from that brand and an edit button. Category buttons are changed similarly. Fix a thread problem in onSwipe.

+++ app/src/main/java/com/canli/oya/traininventory/ui/CategoryListFragment.java
@@ -7,11 +7,13 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.design.widget.CoordinatorLayout;
 import android.support.design.widget.Snackbar;
+import android.support.transition.Slide;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.helper.ItemTouchHelper;
+import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -20,17 +22,19 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.canli.oya.traininventory.R;
 import com.canli.oya.traininventory.adapters.CategoryAdapter;
 import com.canli.oya.traininventory.data.TrainDatabase;
 import com.canli.oya.traininventory.data.entities.CategoryEntry;
 import com.canli.oya.traininventory.utils.AppExecutors;
+import com.canli.oya.traininventory.utils.Constants;
 import com.canli.oya.traininventory.viewmodel.MainViewModel;
 
 import java.util.List;
 
-public class CategoryListFragment extends Fragment {
+public class CategoryListFragment extends Fragment implements CategoryAdapter.CategoryItemClickListener{
 
     private TrainDatabase mDb;
     private TextView empty_tv;
@@ -51,7 +55,7 @@ public class CategoryListFragment extends Fragment {
 
         mDb = TrainDatabase.getInstance(getActivity().getApplicationContext());
 
-        mAdapter = new CategoryAdapter();
+        mAdapter = new CategoryAdapter(this);
         recycler = rootView.findViewById(R.id.list);
         recycler.setLayoutManager(new LinearLayoutManager(getActivity()));
         recycler.setItemAnimator(new DefaultItemAnimator());
@@ -88,32 +92,45 @@ public class CategoryListFragment extends Fragment {
                 final int position = viewHolder.getAdapterPosition();
 
                 //First take a backup of the category to erase
-                final CategoryEntry categoryToErase = new CategoryEntry( mCategories.get(position));
+                final CategoryEntry categoryToErase = new CategoryEntry(mCategories.get(position));
 
                 //Remove the category from the database
                 AppExecutors.getInstance().diskIO().execute(new Runnable() {
                     @Override
                     public void run() {
-                        mDb.categoryDao().deleteCategory(categoryToErase);
+                        //First check whether this category is used by trains table
+                        if (mDb.trainDao().isThisCategoryUsed(categoryToErase.getCategoryName())) {
+                            // If it is used, show a warning and don't let user delete this
+                            getActivity().runOnUiThread(new Runnable() {
+                                @Override
+                                public void run() {
+                                    Toast.makeText(getActivity(), R.string.cannot_erase_category, Toast.LENGTH_LONG).show();
+                                    mAdapter.notifyDataSetChanged();
+                                }
+                            });
+                            return;
+                        } else {
+                            //If it is not used, erase the category
+                            mDb.categoryDao().deleteCategory(categoryToErase);
+                            //Show a snackbar for undoing delete
+                            Snackbar snackbar = Snackbar
+                                    .make(coordinator, R.string.brand_deleted, Snackbar.LENGTH_INDEFINITE)
+                                    .setAction(R.string.undo, new View.OnClickListener() {
+                                        @Override
+                                        public void onClick(View view) {
+                                            //If UNDO is clicked, add the item back in the database
+                                            AppExecutors.getInstance().diskIO().execute(new Runnable() {
+                                                @Override
+                                                public void run() {
+                                                    mDb.categoryDao().insertCategory(categoryToErase);
+                                                }
+                                            });
+                                        }
+                                    });
+                            snackbar.show();
+                        }
                     }
                 });
-
-                //Show a snackbar for undoing delete
-                Snackbar snackbar = Snackbar
-                        .make(coordinator, R.string.brand_deleted, Snackbar.LENGTH_LONG)
-                        .setAction(R.string.undo, new View.OnClickListener() {
-                            @Override
-                            public void onClick(View view) {
-                                //If UNDO is clicked, add the item back in the database
-                                AppExecutors.getInstance().diskIO().execute(new Runnable() {
-                                    @Override
-                                    public void run() {
-                                        mDb.categoryDao().insertCategory(categoryToErase);
-                                    }
-                                });
-                            }
-                        });
-                snackbar.show();
             }
         }).attachToRecyclerView(recycler);
 
@@ -143,13 +160,13 @@ public class CategoryListFragment extends Fragment {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        if(item.getItemId() == R.id.action_add){
+        if (item.getItemId() == R.id.action_add) {
             openAddCategoryFragment();
         }
         return super.onOptionsItemSelected(item);
     }
 
-    private void openAddCategoryFragment(){
+    private void openAddCategoryFragment() {
         AddCategoryFragment addCatFrag = new AddCategoryFragment();
         getChildFragmentManager().beginTransaction()
                 .setCustomAnimations(R.anim.translate_from_top, 0)
@@ -157,4 +174,18 @@ public class CategoryListFragment extends Fragment {
                 .commit();
     }
 
+    @Override
+    public void onCategoryItemClicked(int position) {
+        TrainListFragment trainListFrag = new TrainListFragment();
+        Bundle args = new Bundle();
+        args.putString(Constants.INTENT_REQUEST_CODE, Constants.TRAINS_OF_CATEGORY);
+        args.putString(Constants.CATEGORY_NAME, mCategories.get(position));
+        trainListFrag.setArguments(args);
+        trainListFrag.setEnterTransition(new Slide(Gravity.END));
+        trainListFrag.setExitTransition(new Slide(Gravity.START));
+        getFragmentManager().beginTransaction()
+                .replace(R.id.container, trainListFrag)
+                .addToBackStack(null)
+                .commit();
+    }
 }
